+++ database/migrations/2022_05_24_035625_create_implements_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('implement_model_id')->constrained();
             $table->string('implement_number',5);
-            $table->double('hours');
+            $table->decimal('hours',8,2);
             $table->foreignId('user_id')->constrained();
             $table->timestamps();
             $table->index(['implement_model_id','implement_number']);
