+++ database/migrations/2022_05_24_131152_create_components_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignId('item_id')->unique()->constrained();
             $table->string('component')->unique();
             $table->boolean('is_part');
-            $table->double('lifespan');
+            $table->decimal('lifespan',8,2);
             $table->timestamps();
         });
     }
