+++ database/migrations/2022_05_24_170809_create_tasks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('tasks', function (Blueprint $table) {
             $table->id();
             $table->string('task')->unique();
-            $table->double('estimated_time');
+            $table->decimal('estimated_time',8,2);
             $table->timestamps();
         });
     }
