+++ database/migrations/2022_05_24_172051_create_tractors_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('tractor_model_id')->constrained();
             $table->string('tractor_number',5);
-            $table->double('hour_meter');
+            $table->decimal('hour_meter',8,2);
             $table->timestamps();
             $table->index(['tractor_model_id','tractor_number']);
         });
