+++ database/migrations/2022_05_28_171740_create_min_stocks_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('item_id')->constrained();
             $table->foreignId('warehouse_id')->constrained();
-            $table->double('required_quantity');
-            $table->double('current_quantity');
+            $table->decimal('required_quantity',8,2);
+            $table->decimal('current_quantity',8,2);
             $table->double('price');
             $table->timestamps();
         });
