+++ database/migrations/2022_05_28_172103_create_stockpile_details_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('stockpile_id')->constrained();
             $table->foreignId('item_id')->constrained();
-            $table->double('quantity');
-            $table->double('price');
+            $table->decimal('quantity',8,2);
+            $table->decimal('price',8,2);
             $table->foreignId('warehouse_id')->constrained();
             $table->enum('state',['PENDIENTE','VALIDADO','RECHAZADO','ANULADO']);
             $table->timestamps();
